Move degrees to components
